fixed validation on submit
previously, if one of the validation functions made it to the else statement, there was no return statement, so the function returns undefined which caused the submit button to work when fields weren't filled in.  Added return statement to functions, and if statement to set the value of $error after all functions have been called.